requirements.txt
gspread
oauth2client
pandas
plotly
streamlit==1.32.2
Pillow